Enable handling of version "04" in addition to version "02".
To handle Version "04" we need to parse the valid protocols versions and negotiate the maximally supported one. This sets up the packet size payloads equivalently.

Furthermore, we need to alter the handshake such that we expect the right version of the VtThello message. We simply force correct operation by verification of the version tag.

And finally, VtTread packets may now contain a 4-byte count. We can handle this by extending our binary pattern match accordingly.