Fix protocol version decoding.
Selection of the protocol is done by the client, not by the server. Fix this by letting the client pick the version. The only way this can work is if the first, VtThello packet is `{packet, 2}` encoded. So do that, decode the version, and then pick the packet encoding format for the remaining packets. Reply in the 2 byte packet format and then split off into either 2 or 4 byte encodings.